refactor: Command 인터페이스 패키지 이동(controller -> domain)
